RF: make into meaningful packages, following nilearn/nibabel structure.
